Add recent emojis category
`EmojiDb` now has one more category, "Recent", which is always the first
and contains recently accepted emojis.

To add an emoji to this category, call addRecentEmoji(). This function
places the added emoji at the beginning of the category's emojis and
removes another item being this emoji, if any.

The emojis in the "Recent" category are kept synchronized with user
settings. On Unix, this is in "$HOME/.config/jome/jome.conf". This is an
INI file managed by a `QSettings` instance; it should not be edited
manually. You can remove it if recent emojis are corrupted or something.

In server mode, the "Recent" category is updated immediately when
accepting an emoji so that the category is up to date when showing the
window again.

There's a hard limit of 30 recent emojis as of this patch. This could
become configurable eventually. As a reference point, the iPhone 6S
emoji keyboard (running iOS 12) contains 30 frequently used emojis. The
Messenger app on the same platform also contains 30 frequently used
emojis.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>