Decouple `EmojiDb` from Qt (`QSettings`)
`EmojiDb` is now independent from Qt (it used `QSettings`) and could be
used for other emoji applications along with its required resources.

Instead of the database updating and loading the user settings itself,
everything related to settings is done in `settings.cpp` and the new
EmojiDb::recentEmojis() function sets all the emojis of the database's
recent emojis category.

When the user picks an emoji:

1. Update the database's recent emojis category from the settings (call
   updateRecentEmojisFromSettings()) as the settings might have changed
   in the meantime (while the jome instance's window was visible).

   This reads the settings to build an array of ordered recent emojis
   and calls EmojiDb::recentEmojis() to update the local database.

2. Add the picked emoji to the database's recent emojis as before this
   patch with EmojiDb::addRecentEmoji().

3. Update the settings from the database's recent emojis category
   obtained with EmojiDb::recentEmojisCat() (call updateSettings()).

Before showing the jome window (which is now done with showWindow()),
always update the database's recent emojis category from the settings.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>