Press F12 to go to the Emojipedia page of the selected emoji
This patch adds the F12 keyboard shortcut to open the Emojipedia
page of the selected emoji.

As of this date, you can open the `https://emojipedia.org/XYZ` URL,
where `XYZ` is either the emoji name or the emoji itself, to reach said
page.

jome uses QDesktopServices::openUrl() to open the URL in the user's
browser, not caring about the return value (whether or not it worked).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>