Go to Emojipedia page with context menu
This patch makes it possible for the user to right-click an emoji,
bringing up a context menu, and then click the "Go to Emojipedia page"
action to perform the same action as when pressing F12 with a selected
emoji.

Because this is not configurable, I'm removing the emojiInfoRequested()
signal to embed the URL opening action directly into the main window.
jome::gotoEmojipediaPage(), declared in `jome/emojipedia.hpp`, does this
job.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>