Sort find results by "relevance"
This patch attempts to sort the find results by "relevance".

When finding emojis, there's now a score system for each term:

Exact name match:                                   Set to 100
Partial name match at the beginning:                Set to 80
Partial name match:                                 Set to 60
Exact keyword match (any):                          Add 40
Partial keyword match at the beginning (any):       Add 30
Partial keyword match (any):                        Add 20

The final score is the sum of the partial term scores.

If any term has a score of zero, the emoji remains excluded from the
find results. Changing this would lead to too many results and probably
constitute a UX regression. The find results are automatically sorted in
an `std::set` by score first, and then by global position (this used to
be the only sorting criteria) to break the tie.

The "Recent" category is now excluded. I'm thinking about adding an
option to include it again, but I'll wait to see if someone's angry.

I find that the find results are much more relevant this way, and the
first emoji (the selected one while finding) is often the expected one
with only a few characters in one or two terms.

Fixes: https://github.com/eepp/jome/issues/23
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>