Emoji DB: remove redundant keywords
jome now considers the name when finding emojis, therefore any
keyword which is a substring of the name is redundant.

Leaving a single keyword even if it's redundant (all the flag emojis
have the name `Flag: Xyz` and the keyword `flag` for example) because
it's simpler considering the current assumptions of jome
and `create.py`.

This should theoretically improve find performance because there are
less keywords to consider in EmojiDb::findEmojis().

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>