Find emojis by codepoint
Now you can find emojis by codepoint using the `U+ABCD` notation, for
example `U+1F976` (or `u+1f976`) will find 🥶.

A nice side effect is that you can show all the emojis having a VS-16
codepoint with `u+fe0f` and all the emojis having a ZWJ with `u+200d`.

This is all implemented in EmojiDb::findEmojis(): there needs to be
a single needle, its size needs to be at least three, and it must start
with `u+` (all needles are lowercase).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>